changed the design, reduced the "main.c"
