+++ ProCamp_Task_1/ProCamp_Task_3_RGB888_to_RGB565.c
@@ -1,4 +1,4 @@
-// Task:
+// Task3:
 // Write a function that converts the RGB888 data format to RGB565
 /*
 * RGB888 - 24 bits, 8 bits per component.
@@ -6,15 +6,25 @@
 * https://www.researchgate.net/figure/Bit-rearrangement-substep-applied-to-every-pixel-of-a-source-image-for-obtaining_fig5_333883036
 */
 
-// #include <stdint.h>
 #include "ProCamp_Task_3_RGB888_to_RGB565.h"
 
-uint16_t RGB888_to_RGB565(uint8_t* rgb888Pixel)
+void task3_main_RGB888_to_RGB565()
 {
-    uint8_t red = rgb888Pixel[0];
-    uint8_t green = rgb888Pixel[1];
-    uint8_t blue = rgb888Pixel[2];
+	printf("Task 3: RGB8888 to RGB565 \n");
+	printf("Start \n\n");
 
+	uint8_t rgb888Pixel[3] = { 0b10010001, 0b10011001, 0b11111111 };
+	uint16_t rgb565Pixel = RGB888_to_RGB565(rgb888Pixel[0], rgb888Pixel[1], rgb888Pixel[2]);
+
+	print_rgb888Pixel(rgb888Pixel);
+	print_rgb565Pixel(rgb565Pixel);
+
+	printf("Task 3: RGB8888 to RGB565 \n");
+	printf("Stop \n\n");
+}
+
+uint16_t RGB888_to_RGB565(uint8_t red, uint8_t green, uint8_t blue)
+{
     uint16_t b = (blue >> 3) & 0x1f;
     uint16_t g = ((green >> 2) & 0x3f) << 5;
     uint16_t r = ((red >> 3) & 0x1f) << 11;
@@ -22,12 +32,28 @@ uint16_t RGB888_to_RGB565(uint8_t* rgb888Pixel)
     return (uint16_t)(r | g | b);
 }
 
-// TODO:
-//uint32_t RGB565_to_RGB888(uint16_t rgb565)
-//{
-//    uint32_t B = (uint32_t)(((rgb565 & 0x001F) << 3) | (rgb565 & 0x0007));
-//    uint32_t G = (uint32_t)(((rgb565 & 0x07E0) << 5) | ((rgb565 & 0x0060) << 3));
-//    uint32_t R = (uint32_t)(((rgb565 & 0xF800) << 8) | ((rgb565 & 0x3800) << 5));
-//
-//    return (R | G | B);
-//}
+void print_binary(unsigned int number)
+{
+	if (number >> 1) {
+		print_binary(number >> 1);
+	}
+	putc((number & 1) ? '1' : '0', stdout);
+}
+
+void print_rgb888Pixel(uint8_t* rgb888Pixel)
+{
+	printf("uint8_t rgb888Pixel[3] = RGB(8+8+8=24) \n\t\t\t");
+	print_binary(rgb888Pixel[0]);
+	printf(" ");
+	print_binary(rgb888Pixel[1]);
+	printf(" ");
+	print_binary(rgb888Pixel[2]);
+	printf("\n");
+}
+
+void print_rgb565Pixel(uint16_t rgb565Pixel)
+{
+	printf("uint16_t rgb565Pixel = RGB(5+6+5=16) \n\t\t\t");
+	print_binary(rgb565Pixel);
+	printf("\n\n");
+}
