+++ ProCamp_Task_1/ProCamp_Task_6_factorial.c
@@ -1,8 +1,21 @@
-/*
-* Write a program that computes the factorial of 10
-*/
+// Task6:
+// Write a program that computes the factorial of 10
 
 // factorial of n (n!) = 1 * 2 * 3 * 4....n
+
+#include "ProCamp_Task_6_factorial.h"
+
+void task6_main_factorial()
+{
+    printf("Task 6: the factorial of 10 \n");
+    printf("Start \n\n");
+
+    int n = 10;
+    printf("Factorial of %d = %ld \n\n", n, multiplyNumbers(n));
+
+    printf("Task 6: the factorial of 10 \n");
+    printf("Stop \n\n");
+}
 
 long int multiplyNumbers(int n)
 {
@@ -19,11 +32,3 @@ long int multiplyNumbers(int n)
         return fact;
     }
 }
-
-//long int multiplyNumbers(int n) 
-//{
-//    if (n >= 1)
-//        return n * multiplyNumbers(n - 1);
-//    else
-//        return 1;
-//}
