+++ ProCamp_Task_1/ProCamp_Task_6_factorial.h
@@ -1,3 +1,15 @@
+// Task6:
+// Write a program that computes the factorial of 10
+/*
+* tvoid task6_main_factorial(void); // main() for task6
+*
+* long int multiplyNumbers(int n); // find the factorial of n
+*/
+
 #pragma once
+
+#include <stdio.h>
+
+void task6_main_factorial(void);
 
 long int multiplyNumbers(int);
