+++ ProCamp_Task_1/ProCamp_Task_7_divisible.h
@@ -1,3 +1,18 @@
+// Task7:
+// Write a program that prints all the integer numbers between 1 to
+// 500 which are divisible by a number received from a user
+/*
+* void task7_main_divisible(void); // main() for task7
+*
+* int divisible(int temp, int n, int i); // temp = 500, n = a number received from a user, i = counter; return n * i which is less than 500 (temp)
+*/
+
 #pragma once
+
+#define _CRT_SECURE_NO_WARNINGS
+
+#include <stdio.h>
+
+void task7_main_divisible(void);
 
 int divisible(int, int, int);
