+++ ProCamp_Task_1/main.h
@@ -6,18 +6,3 @@
 #include "ProCamp_Task_5_circle.h"
 #include "ProCamp_Task_6_factorial.h"
 #include "ProCamp_Task_7_divisible.h"
-
-void task2_main_swap(void);
-
-void task3_main_RGB888_to_RGB565(void);
-void print_binary(unsigned int);
-void print_rgb888Pixel(uint8_t*);
-void print_rgb565(uint16_t);
-
-void task4_main_inches(void);
-
-void task5_main_circle(void);
-
-void task6_main_factorial(void);
-
-void task7_main_divisible(void);
